+++ src/stdlibs/ccalloc.c
@@ -5,9 +5,11 @@
 ** malloc n strings and fill with 0
 */
 
+#include <stddef.h>
 #include <stdlib.h>
+#include "tlcstdlibs.h"
 
-char **x_ccalloc(int n)
+char **x_ccalloc(size_t n)
 {
     char **new = NULL;
 
@@ -15,7 +17,7 @@ char **x_ccalloc(int n)
     if (new == NULL) {
         return (NULL);
     }
-    for (int i = 0; i < n; i++) {
+    for (size_t i = 0; i < n; i++) {
         new[i] = NULL;
     }
     return (new);
