+++ src/strings/strlen.c
@@ -7,13 +7,14 @@
 
 #include <stddef.h>
 #include <stdbool.h>
+#include "tlcstrings.h"
 
 size_t x_strlen(const char *s)
 {
     if (s == NULL) {
         return (0);
     }
-    for (int i = 0; true; i++) {
+    for (size_t i = 0; true; i++) {
         if (s[i] == 0) {
             return (i);
         }
